pyproject.toml
[project]
name = "apl-calculator"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Kingston",email = "71872141+mkingopng@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "constructs (>=10.4.2,<11.0.0)",
    "aws-cdk-lib (>=2.177.0,<3.0.0)",
    "flask-lambda (>=0.0.4,<0.0.5)",
    "aws-lambda-powertools (>=3.5.0,<4.0.0)",
    "boto3 (>=1.36.8,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "mangum (>=0.19.0,<0.20.0)",
    "pytest (>=8.3.4,<9.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "flask_app" }]
